requirements.txt
atomicwrites==1.4.1
attrs==22.1.0
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
idna==3.3
iniconfig==1.1.1
jmespath==1.0.1
packaging==21.3
pathlib==1.0.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
pytz==2022.1
pywin32==304
requests==2.28.1
simplejson==3.17.6
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
toml==0.10.2
tomli==2.0.1
traits==6.3.2
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
XlsxWriter==3.0.3
